# swaggerDoc.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Biblioteca",
    "description": "API para gerenciamento de usuários, livros e movimentações em uma biblioteca",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/livros/": {
      "get": {
        "tags": [
          "Livros"
        ],
        "description": "Endpoint para retornar todos os livros",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Livros encontrados com sucesso"
          },
          "401": {
            "description": "Token não informado ou inválido"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Livros"
        ],
        "description": "Endpoint para criar um livro",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "livro",
            "in": "body",
            "description": "Informações do livro",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "titulo": {
                  "type": "string",
                  "example": "O Senhor dos Anéis"
                },
                "genero": {
                  "type": "string",
                  "example": "Fantasia"
                },
                "autor": {
                  "type": "string",
                  "example": "J. R. R. Tolkien"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Livro criado com sucesso"
          },
          "401": {
            "description": "Token não informado ou inválido"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/livros/{id}": {
      "get": {
        "tags": [
          "Livros"
        ],
        "description": "Endpoint para buscar um livro por id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Id do livro"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Livro encontrado com sucesso"
          },
          "401": {
            "description": "Token não informado ou inválido"
          },
          "404": {
            "description": "Livro não encontrado"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Livros"
        ],
        "description": "Endpoint para atualizar um livro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Id do livro"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "livro",
            "in": "body",
            "description": "Informações do livro",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "titulo": {
                  "type": "string",
                  "example": "O Senhor dos Anéis"
                },
                "genero": {
                  "type": "string",
                  "example": "Fantasia"
                },
                "autor": {
                  "type": "string",
                  "example": "J. R. R. Tolkien"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Livro atualizado com sucesso"
          },
          "401": {
            "description": "Token não informado ou inválido"
          },
          "404": {
            "description": "Livro não encontrado"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Livros"
        ],
        "description": "Endpoint para deletar um livro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Id do livro"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Livro deletado com sucesso"
          },
          "401": {
            "description": "Token não informado ou inválido"
          },
          "404": {
            "description": "Livro não encontrado"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  }
}